Reorganize a little
