+++ libp2p-peerstore/src/lib.rs
@@ -4,3 +4,6 @@ extern crate libp2p_peer as peer;
 
 mod memory_peerstore;
 mod peerstore;
+mod peer_info;
+
+pub type TTL = std::time::Duration;
